Add Python script: Sort colors in picture.
The left upper corner of the picture is designated as the red point.
The right upper corner of the picture is designated as the green point.
The lower middle point of the picture is designated as the blue point.
Each pixel is gravitated to the three points according to its RGB
values. After a position is found on which the pixel would land after
its attraction, it is swapped with the pixel in that position.