[rust] Ackermann: fixup! Track unresolved terms within the stack
I will probably replace this with a separate list of terms that are
either unresolved or resolved, where unresolved terms are associated
with the index on the stack they're waiting for;

stack    terms (hashmap)
2 1      [2,1?0]
1 2 0    [2,1?0 2,0?1]
1 1 1    [2,1?0 2,0?1 1,1?1]
1 0 1 0  [2,1?0 2,0?1 1,1?1 1,0?2]
1 0 0 1  [2,1?0 2,0?1 1,1?1 1,0=2]
1 0 2    [2,1?0 2,0=3 1,1=3 1,0=2 1,3?0]
1 3      [2,1?0 2,0=3 1,1=3 1,0=2 1,3?0]
0 1 2    [2,1?0 2,0=3 1,1=3 1,0=2 1,3?0 1,2?1]
...

However, this approach has a drawback: While moving back
up (i.e. shrinking) the stack we need to iterate through all values of
the hashmap to see whether the index is in an unresolved term.

More efficient would be to track unresolved terms not in the hashmap,
but in a separate list, since we could simply use the stack index as
lookup for that list.[1]

So while growing the stack we'll look up (m,n) in the hashmap of
resolved term, and while shrinking the stack we'll look up the stack
index in the list of unresolved terms to see whether we can resolve any.

stack    unresolved terms (list)          resolved terms (hashmap)
2 1      [(2 1)]
1 2 0    [(2 1) (2 0)]
1 1 1    [[(2 1)] [(2 0) (1 1)]]
1 0 1 0  [[(2 1)] [(2 0) (1 1)] [(1 0)]]
1 0 0 1  [[(2 1)] [(2 0) (1 1)] [(1 0)]]
1 0 2    [[(2 1)] [(2 0) (1 1)]]          [1,0=2]
1 3      [[(2 1) (1 3)]]                  [1,0=2 2,0=3 1,1=3]
0 1 2    [[(2 1) (1 3)] [(1 2)]]          [1,0=2 2,0=3 1,1=3]
0 0 1 1  [[(2 1) (1 3)] [(1 2)]]          [1,0=2 2,0=3 1,1=3]
0 0 3    [[(2 1) (1 3)] [(1 2)]]          [1,0=2 2,0=3 1,1=3]
0 4      [[(2 1) (1 3)]]                  [1,0=2 2,0=3 1,1=3 1,2=4]
5        []                               [1,0=2 2,0=3 1,1=3 1,2=4 2,1=5 1,3=5]

[1] It's okay that the list of unresolved terms only tracks their stack
index implicitly (i.e. via its own index into the list), because we
can guarantee that no stack index will be skipped;
- Either the term at a stack index is instantly resolvable (m==0) so
  we go back up the stack...
- or it isn't, in which case it's either
  - a term that we know and thus can instantly resolve (going back up
    the stack)
  - or a term we have to add to the list of unresolved terms.