fixup! Test for descrete uniform distribution of N-bit values after sha256 'reroll'
Let's say I want to deterministically choose a pseudo-random number
between 1 and N (including) in a way that makes each number equally
likely (this is called a descrete uniform distribution), but N is not
a power of 2, so I cannot simply take the corresponding amount of bits
from a random bit stream.

I could take p bits where 2^(p-1)<N and 2^p>N and take the result
modulo N, but then the first 2^p mod N numbers would have a higher
probability than the others.

Therefore I need to take p bits and reroll until the resulting encoded
unsigned integer number is within 1 and N (including).

This should make the probability of each number equal. But I want to
test this empirically. This commit adds that empirical test for 1<=p<=8.